requirements.txt
flask>=2.3.2
flask-socketio>=5.3.4
python-socketio>=5.8.0
python-engineio>=4.5.1
Pillow>=10.0.0
pytesseract>=0.3.10
numpy>=1.24.0
opencv-python>=4.8.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0